Simplify AssetsPublishCommand
