+++ src/Pug/Symfony/Traits/PrivatePropertyAccessor.php
@@ -2,10 +2,22 @@
 
 namespace Pug\Symfony\Traits;
 
+use ReflectionException;
 use ReflectionProperty;
 
 trait PrivatePropertyAccessor
 {
+    /**
+     * Expose a private/protected property.
+     *
+     * @param object             $object
+     * @param string             $property
+     * @param ReflectionProperty $propertyAccessor
+     *
+     * @throws ReflectionException
+     *
+     * @return mixed
+     */
     public static function getPrivateProperty($object, $property, &$propertyAccessor = null)
     {
         $propertyAccessor = new ReflectionProperty($object, $property);
